save items to local

+++ TodoListApp/View/ListView.swift
@@ -12,20 +12,26 @@ struct ListView: View {
     @EnvironmentObject var listViewModel: ListViewModel
     
     var body: some View {
-        List {
-            ForEach(listViewModel.items){item in
-                ListRowView(item: item)
-                    .onTapGesture {
-                        withAnimation(.linear){
-                            listViewModel.updateItem(item: item)
-                        }
+        ZStack {
+            if listViewModel.items.isEmpty{
+                Text("No items 🙁")
+            } else {
+                List {
+                    ForEach(listViewModel.items){item in
+                        ListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear){
+                                    listViewModel.updateItem(item: item)
+                                }
+                            }
                     }
+                    .onMove(perform: listViewModel.moveItem)
+                    .onDelete(perform: listViewModel.deleteItem)
+                }
+                
+                .listStyle(PlainListStyle())
             }
-            .onMove(perform: listViewModel.moveItem)
-            .onDelete(perform: listViewModel.deleteItem)
         }
-        
-        .listStyle(PlainListStyle())
         .navigationTitle("Todo List 📝")
         .navigationBarItems(
             leading: EditButton(),
